Fix backprop for broadcasting + 100% test coverage
Fixed the bug that used to happen during backpropagation if an
operation, like addition, multiplication, etc., required broadcasting to
be fulfilled. In such cases, when we do the backward phase, we need to
make sure to accumulate gradients from all dimensions where broadcasting
happened.

I also increased test coverage to 100%, and set the test threshold to
99% to ensure that new code is covered by tests.